docker/requirements.txt
requests
smllib
paho-mqtt
